+++ generator/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 8.1.0
+
+- Added `@Extras` to pass extra options to dio requests, response, transformer and interceptors.
+
+  Example :
+  ```dart
+  @http.POST('/path/')
+  Future<String> myMethod(@Extras() Map<String, dynamic> extras);
+  ```
+
+## 8.0.6
+
+- @useResult
+
+## 8.0.5
+
+- fix [#627](https://github.com/trevorwang/retrofit.dart/issues/627) where generic argument constructors on any class not directly marked with `@JsonSerializable(genericArgumentFactories: true)`
+
+- fix some typos in this changelog.
+
 ## 8.0.2
 
 - fix #630 Null check operator used on a null value
@@ -123,7 +143,7 @@
 
 ## 2.0.0-beta1
 
-- Nullsafety support
+- Null safety support
 
 ## 1.4.1
 
@@ -135,7 +155,7 @@
 
 ## 1.3.8
 
-- Send list params duplicative in multipart
+- Send list params duplication in multipart
 
 ## 1.3.7
 
@@ -190,7 +210,7 @@
 
 ## 1.2.0
 
-- Add `HttpReposne` to handle the original response
+- Add `HttpResponse` to handle the original response
 
 ## 1.1.0
 
@@ -253,11 +273,11 @@ Added bean class support for `@Body()` annotation.
 
 Here's the example.
 
-```
+```dart
     Future<String> createUser(@Body() User user);
 ```
 
-```
+```dart
     class User {
         Map<String, dynamic> toJson() => {};
     }